RCM_ai_backend/requirements.txt
django
django-rest-framework
django-cors-headers
transformers
crewai
langchain
langchain-openai
python-dotenv
pandas
numpy
scikit-learn
spacy
pytesseract
pillow
fhir.resources
hl7apy
chromadb
langchain-community
sentence-transformers
pypdf
langchain-chroma